chore(gifts): move codigo para gifts app
